move fixtures in a dedicated directory with a file per type of data
