+++ backend/geonature/tests/fixtures/__init__.py
@@ -1,0 +1,3 @@
+from .general import *
+from .general import _app, _session
+from .habitat import *
